requirements.txt
linkedin-scraper
python-dotenv
selenium